requirements.txt
langchain-community
langchain-text-splitters
langchain-groq
langchain-mistralai
langchain-core
streamlit
python-dotenv
faiss-cpu